requirements.txt
asgiref==3.7.2
cloudinary==1.38.0
dj-rest-auth==2.1.9
Django==3.2.24
django-allauth==0.44.0
django-cloudinary-storage==0.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
oauthlib==3.2.2
pillow==10.2.0
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2024.1
requests-oauthlib==1.3.1
sqlparse==0.4.4
